pyproject.toml
[project]
name = "boilerplate-python-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "Tafarel Yan",email = "tafarel.yan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "boto3 (>=1.40.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
packages = [
    { include = "src" }
]
